Added files via upload
I have made 3 accounts (1 for each of us) the sign in is based on firstname@lastname.com and password is firstname. I hope I spelt your last name right Elena I spelt it "Tsanacheva"?
There aren't any images with the upload as it will be a different file name each time and I didn't want to include files in the commit.